# docs/features/product-creation-flow/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: product-creation-flow API
  version: 0.1.0
  description: |
    Картка комісійного товару як єдине місце, де живуть тексти під Prom і OLX, ціна,
    галерея до десяти кадрів і дві незалежні відмітки присутності. Розрив між
    генерацією тексту в зовнішньому чаті й карткою товару (idea-brief §2) закривається
    тим, що підготовка й ручне редагування пишуть в ті самі поля тим самим маршрутом
    (AC-12) — окремого "прийняти як AI-підготовлене" немає.

    **Поставка 1** (у цій версії контракту — реалізована): картка, галерея, зовнішнє
    сховище, відмітки присутності. Ендпоінти `/products`, `/products/{id}`,
    `/products/{id}/images*`.

    **Поставка 2** (спроектована в data-model.md, таблиці створені міграцією): тексти,
    діапазон ціни, пропозиції, вартість підготовки. Ендпоінти
    `/products/{id}/preparation-runs*` і `/products/{id}/suggestions/*` описані тут
    наперед, бо форма зберігання вже зафіксована ([sad.md §1](../sad.md)), а розділяти
    контракт на дві версії означало б переписувати схему `Product` двічі.

    Джерела: [PRD](../PRD.md) · [data-model.md](../data-model.md) · [sad.md §6](../sad.md).
    Аудит розбіжностей — [api-sync-report.md](./api-sync-report.md).
servers:
  - url: https://mouse.springconsult.com.ua/api
    description: Production (Caddy handle_path /api/* → api:3000, префікс знято)
  - url: http://localhost:4200/api
    description: Local dev (ng serve proxy.conf.json, той самий префікс)
security:
  - sessionCookie: []
...
paths:
...
  /products/{productId}/preparation-runs:
    parameters:
      - { $ref: "#/components/parameters/ProductId" }

    get:
      tags: [preparation]
      summary: Невдалі запуски картки
      description: >-
        Список для попапу каталогу ([T50](../tasks/show-preparation-failures-in-catalog.md), AC-49).
        Поки що вміє лише `status=failed`: інших фільтрів каталог не просить. Найновіші
        першими; без пагінації, бо ліміт частоти запусків тримає їх кількість малою.
      operationId: listFailedPreparationRuns
      parameters:
        - name: status
          in: query
          required: true
          schema: { type: string, enum: [failed] }
      responses:
        "200":
          description: Невдалі запуски картки, порожній масив, якщо їх не було
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/PreparationRun" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/NotAuthenticated" }
        "404": { $ref: "#/components/responses/ProductNotFound" }

    post:
      tags: [preparation]
      summary: Запустити підготовку (тексти, ціна або обидва)
      description: >-
        **Поставка 2 — ендпоінт спроектований, таблиця `product_preparation_runs`
        створена міграцією (data-model.md, розділ Migrations).**

        Область `texts`/`both` розпізнає товар із головного кадру автоматично, як
        частину того самого виклику, що готує тексти ([ADR 0014](../adr/0014-let-ai-recognize-the-item-from-photos.md));
        область `field` регенерує одне поле з чернетки в тілі запиту, без фото
        ([ADR 0015](../adr/0015-add-per-field-text-rewrite-scope.md)).

        Гейт AC-06 ("немає жодного кадру") перевіряється синхронно до постановки
        задачі — перевірка коштує одного читання, запуск коштує грошей (sad.md
        сценарій 7). Ідемпотентність — **не** заголовок клієнта:
        ключ (картка, область, версія входу) обчислює сервер із `idempotency_key`
        UNIQUE у `product_preparation_runs`; повторний виклик з тим самим входом
        повертає наявний запуск (`200`), а не створює новий (`201`). Наявним є лише
        незавершений (`queued`, `running`) чи `succeeded` запуск: UNIQUE частковий,
        `WHERE status <> 'failed'`, бо відмовлений запуск результату не має й захищати
        в ньому нема чого. Після `failed` той самий вхід дає `201` і нову задачу
        (AC-37). "Версія входу" — відкритий пункт data-model.md (Open items),
        контракт лишає його явним, а не ховає.
      operationId: startPreparationRun
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PreparationRunCreateRequest" }
            example: { scope: both }
      responses:
        "201":
          description: Новий запуск поставлено в чергу
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PreparationRun" }
        "200":
          description: Той самий вхід уже має незавершений запуск — повернуто його
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PreparationRun" }
        "401": { $ref: "#/components/responses/NotAuthenticated" }
        "404": { $ref: "#/components/responses/ProductNotFound" }
        "409":
          description: >-
            У галереї немає жодного кадру для scope: texts/both (AC-06), або в
            картці немає ні titleProm, ні titleOlx для scope: price (AC-27) —
            два різні missing під тим самим кодом
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              examples:
                gallery:
                  value: { error: { code: preparation_input_incomplete, message: "Додайте бодай один кадр перед підготовкою", details: { missing: ["gallery"] } } }
                title:
                  value: { error: { code: preparation_input_incomplete, message: "Додайте заголовок Prom або OLX перед пошуком ціни", details: { missing: ["title"] } } }
        "429":
          description: Перевищено обмеження частоти запусків — 20 на картку за годину, спільні для всіх областей (PRD §6.1 — окреме від ліміту входу)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              example:
                error: { code: preparation_rate_limited, message: "Забагато запусків підряд, спробуйте пізніше" }
  /products/{productId}/preparation-runs/{runId}:
    parameters:
      - { $ref: "#/components/parameters/ProductId" }
      - { $ref: "#/components/parameters/RunId" }

    get:
      tags: [preparation]
      summary: Стан запуску (полінг)
      description: >-
        **Поставка 2 — спроектовано.** sad.md сценарії 7-8: фронт питає стан у циклі,
        поки запуск не завершено; `errorCode` заповнюється лише при `status: failed`
        і є тим самим доменним кодом, який AC-10 показує людині.
      operationId: getPreparationRun
      responses:
        "200":
          description: Стан запуску
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PreparationRun" }
        "401": { $ref: "#/components/responses/NotAuthenticated" }
        "404":
          description: Картку або запуск не знайдено
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: mouse_session
      description: >-
        JWT (HS256, jose) у HttpOnly-cookie ([ADR 0002](../../../adr/0002-auth-jwt-and-typeorm.md)
        · `modules/auth/CLAUDE.md`). Не Bearer-токен — SPA й API на одному домені,
        токен не має причин бути доступним для JavaScript. Ролей немає, перевірка
        одна: активна сесія (`sessionGuard`).
  parameters:
    ProductId:
      name: productId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    RunId:
      name: runId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      description: >-
        Обгорнуто в `error` (не пласко `{code, message}`) — форма наявного
        `contracts/error.contract.ts`, скіл-дефолт узгоджено з нею, а не навпаки.
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: gallery_full }
            message: { type: string }
            details: { type: object }
...
    PreparationRunCreateRequest:
      type: object
      required: [scope]
      description: >-
        `field` і `draftText` обов'язкові лише при `scope: field` ([ADR 0015](../adr/0015-add-per-field-text-rewrite-scope.md));
        для решти областей вони відсутні. Валідує це `z.discriminatedUnion` на межі
        контролера, не ця схема.
      properties:
        scope:
          type: string
          enum: [texts, price, both, field]
          description: >-
            "ціна" без перезапуску текстів — AC-10b має предмет саме завдяки цій
            області (data-model.md, product_preparation_runs.scope). "texts"/"both"
            розпізнають товар із головного кадру автоматично ([ADR 0014](../adr/0014-let-ai-recognize-the-item-from-photos.md));
            "field" регенерує одне поле з `draftText`, без фото ([ADR 0015](../adr/0015-add-per-field-text-rewrite-scope.md)).
        field:
          type: string
          enum: [titleProm, titleOlx, descriptionProm, descriptionOlx, seoKeywords]
          description: "Лише при `scope: field` — яке поле регенерувати."
        draftText:
          type: string
          description: >-
            Лише при `scope: field` — поточна незбережена чернетка поля, з якої
            модель готує новий варіант.
      example: { scope: both }
    PreparationRun:
      type: object
      description: "**Поставка 2 — спроектовано, таблиця створена міграцією.**"
      required: [id, productId, scope, status, model, inputTokens, outputTokens, createdAt]
      properties:
        id: { type: string, format: uuid }
        productId: { type: string, format: uuid }
        scope: { type: string, enum: [texts, price, both, field] }
        status: { type: string, enum: [queued, running, succeeded, failed] }
        errorCode:
          type: [string, null]
          description: >-
            Заповнено лише при status=failed; той самий код, що показує AC-10.
            `price_unavailable` — часткова відмова `scope: both`: пропозиції текстів
            записані, пропозиції `price` немає, ціну просить окремий `scope: price` (AC-10b).
            `preparation_failed` — усі спроби задачі впали, пропозицій немає (AC-10, events.md).
        errorDetail:
          type: [string, null]
          description: >-
            Лише при status=failed: англійське повідомлення помилки, що закрила запуск, до
            500 символів. Технічна деталь під українським текстом за `errorCode`, не заміна
            йому. `null` у запусків, що впали до T50.
        model: { type: string, example: claude-sonnet-5 }
        inputTokens: { type: integer, minimum: 0 }
        outputTokens: { type: integer, minimum: 0 }
        createdAt: { type: string, format: date-time }
        startedAt: { type: [string, null], format: date-time }
        finishedAt: { type: [string, null], format: date-time }
...
  responses:
    ValidationError:
      description: Тіло запиту не пройшло zod-валідацію
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          example:
            error:
              code: validation_failed
              message: "Query parameters are invalid"
              details: { fields: { price: ["Price must be a decimal such as 2499.00"] } }
    NotAuthenticated:
      description: Немає активної сесії
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          example:
            error: { code: not_authenticated, message: "Authentication required" }
    ProductNotFound:
      description: Картку не знайдено
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          example:
            error: { code: product_not_found, message: "Product not found" }
...